unified_requirements.txt
langchain>=0.1.0
langgraph>=0.0.50
langchain_community>=0.0.20
langchain_core>=0.1.0
langchain_groq>=0.1.0
langchain_openai>=0.1.0
streamlit>=1.28.0
python-dotenv>=1.0.0
pydantic>=2.0.0
tavily-python>=0.3.0
faiss-cpu>=1.7.4beautifulsoup4>=4.12.0
requests>=2.31.0
tf-keras